+++ CHANGELOG.md
@@ -14,7 +14,10 @@
 - Log entries for everything surrounding the PDFs are stored in the DB
 - Also logs language and course API calls for statistics
 - To track what the user changes between renders, the server returns the id of the logged PDF. The frontend then sends that ID in the request of an eventual new PDF
+
+#### Stats
 - A `/stats` API endpoint (protected by auth) which collates logs from the DB
+- Leaderboard endpoints for topic inclusion and problem exclusion
 
 ### Changed 
 - Moved common styling elements to the `App.css` file